+++ packages/common/src/index.ts
@@ -1,2 +1,3 @@
 export * from './consts';
 export * from './utils';
+export * from './substrate';