+++ packages/xc-core/src/index.ts
@@ -3,5 +3,4 @@ export * from './bridge';
 export * from './chain';
 export * from './config';
 export * from './evm';
-export * from './substrate';
 export * from './utils';