+++ packages/xc-sdk/src/Wallet.ts
@@ -1,7 +1,6 @@
-import { big } from '@galacticcouncil/common';
+import { acc, big } from '@galacticcouncil/common';
 
 import {
-  acc,
   addr,
   Asset,
   AnyChain,